refactoring full TDD
