+++ __old/examples/bads/NotGoodHandlerCommand.js
