+++ __old/examples/client-publish.js
