+++ __old/examples/integration/BasicNopeRejectCommand.js
