+++ __old/examples/integration/BasicNopeResolveCommand.js
