+++ __old/examples/queries/BasicNopeQuery.js
