+++ __old/test/mocks/Rabbot.js
