+++ __old/test/unit/001-errors.test.js
@@ -3,6 +3,7 @@
 const Server = require('../..').Server
 const Client = require('../..').Client
 
+const path = require('path')
 const chai = require('chai')
 const expect = chai.expect
 
@@ -12,10 +13,13 @@ const rabbotMock = new RabbotMock()
 describe('[unit] errors', function () {
   it('should be catch on server without valid rabbitmq connection', function (done) {
     const server = new Server({
-      rabbot: rabbotMock,
       bus: {
         port: 1111
-      }
+      },
+      source: path.resolve(__dirname, '../..', 'examples'),
+      patterns: [
+        'integration/**/*.js'
+      ]
     })
     server.initialize()
       .then(() => {
@@ -33,13 +37,15 @@ describe('[unit] errors', function () {
         done()
       })
   })
-  it('should be catch on server without files to initialize CQRS', function (done) {
+  xit('should be catch on server without files to initialize CQRS', function (done) {
     const server = new Server({
-      rabbot: rabbotMock,
       bus: {
         port: 6666
       },
-      source: 'no_files_here_to_initialize'
+      source: path.resolve(__dirname, '../..', 'examples'),
+      patterns: [
+        'nope/**/*.js'
+      ]
     })
     server.initialize()
       .then(() => {
@@ -58,12 +64,7 @@ describe('[unit] errors', function () {
       })
   })
   it('should be catch on server without options on initialize', function (done) {
-    const server = new Server({
-      rabbot: rabbotMock,
-      bus: {
-        port: 1111
-      }
-    })
+    const server = new Server()
     server.initialize()
       .then(() => {
         done(new Error('no error detected'))
@@ -81,12 +82,13 @@ describe('[unit] errors', function () {
       })
   })
   it('should be catch on client without rabbitmq connection', function (done) {
-    const client = new Client({
-      rabbot: rabbotMock,
+    const options = {
       bus: {
         port: 1111
-      }
-    })
+      },
+      rabbot: rabbotMock
+    }
+    const client = new Client(options.bus, options.rabbot)
     client.initialize()
       .then(() => {
         done(new Error('no error detected'))
