+++ lib/utils/Service.js
@@ -1,0 +1,4 @@
+class Service {
+}
+
+module.exports = Service
