+++ package.json
@@ -6,9 +6,7 @@
   "scripts": {
     "test": "./npm-scripts/test.sh",
     "test:standard": "node_modules/.bin/standard --verbose | snazzy",
-    "test:unit": "node_modules/.bin/istanbul cover node_modules/.bin/_mocha -- --timeout 5000 --use_strict --recursive test/unit",
-    "test:integration": "node_modules/.bin/istanbul cover node_modules/.bin/_mocha -- --timeout 5000 --use_strict --recursive test/integration",
-    "test:coverage": "node_modules/.bin/istanbul cover -x 'examples/**/*.*' node_modules/.bin/_mocha -- --opts .mocharc"
+    "test:coverage": "./node_modules/.bin/nyc --reporter=lcov --reporter=text ./node_modules/.bin/mocha --opts .mocharc"
   },
   "repository": {
     "type": "git",
@@ -28,7 +26,7 @@
     "bluebird": "3.5.1",
     "debug": "3.1.0",
     "eraro": "0.4.1",
-    "glob": "7.1.2",
+    "glob-promise": "3.3.0",
     "hoek": "5.0.2",
     "lodash": "4.17.4",
     "rabbot": "1.1.0",
@@ -38,16 +36,17 @@
   "devDependencies": {
     "chai": "4.1.2",
     "chance": "1.0.12",
-    "istanbul": "0.4.5",
     "joi": "13.0.2",
     "mocha": "4.0.1",
     "mocha-standard": "1.0.0",
+    "nyc": "11.3.0",
     "sinon": "4.1.2",
     "snazzy": "7.0.0",
     "standard": "10.0.3"
   },
   "standard": {
     "globals": [
+      "xit",
       "beforeEach",
       "describe",
       "it"
