--- test/coverage-max.test.js
@@ -1,7 +1,0 @@
-'use strict'
-
-describe('[integration] coverage maximum', function () {
-  it('should ...', function (done) {
-    done()
-  })
-})
