+++ test/data/queries/BasicNopeQuery.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const Promise = require('bluebird')
+
+const handler = function (params) {
+  return new Promise((resolve, reject) => {
+    resolve(params)
+  })
+}
+
+module.exports = handler
